Move util functions inside of `mutation.py`.
